requirements.txt
streamlit
pytesseract
pdf2image
openai
PyMuPDF
Pillow
python-docx